Factored geometry creation into another Slicer CLI module
This portion of the geometry pipeline takes a while to
run. If several attempts are required to find the right
parameters for the segmentation algorithm, the geometry
generation wastes significant amounts of time. You can still
optionally generate the geometry in the AirwaySegmenter
CLI, but if you chose not to during the binary segmentation
phase, you can generate the geometry from the resampled
image, the binary segmentation, and the threshold value
determined by the segmentation algorithm.